LayoutTests/platform/chromium/permissionclient/image-permissions-expected.txt

PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif): true
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif): true
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif?nocache): false
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif?nocache): false
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif): false
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif): false
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif): false
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif): false
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif): false
PERMISSION CLIENT: allowImage((file test):platform/chromium/permissionclient/resources/boston.gif): false


PASS: first image loaded
